back_2.1/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
bcrypt==4.0
passlib[bcrypt]
python-jose[cryptography]
pytest
celery==5.3.6
celery[redis]